refactor: slack block command structure
