+++ server/slack/commands/ping.ts
@@ -4,6 +4,15 @@ import type {
 } from '@slack/bolt';
 import { Actions, Button, Section } from 'slack-block-builder';
 import { asBlocks } from '~/lib/slack/blocks';
+import type { CommandHelp } from '~/types';
+
+export const help: CommandHelp = {
+  name: 'ping',
+  description: 'Check if Gork is alive.',
+  subcommands: [
+    { usage: 'ping', description: 'Responds with pong if Gork is online.' },
+  ],
+};
 
 export async function execute({
   ack,
