+++ server/slack/features/mode/commands/index.ts
@@ -2,7 +2,45 @@ import type {
   AllMiddlewareArgs,
   SlackCommandMiddlewareArgs,
 } from '@slack/bolt';
+import type { CommandHelp } from '~/types';
 import { parseArgs } from './args';
+
+export const help: CommandHelp = {
+  name: 'mode',
+  description: 'Control when Gork replies in this workspace or channel.',
+  subcommands: [
+    {
+      usage: 'mode set [workspace|channel] <mode>',
+      description:
+        'Set the reply mode. Omit scope to open a modal. Workspace scope requires admin.',
+    },
+    {
+      usage: 'mode show [workspace|channel]',
+      description: 'Show stored modes and the effective mode for this channel.',
+    },
+    {
+      usage: 'mode clear <workspace|channel>',
+      description: 'Clear a stored mode. Workspace scope requires admin.',
+    },
+  ],
+  modes: [
+    { name: 'ping', description: 'Only respond when directly @mentioned.' },
+    {
+      name: 'relevance',
+      description:
+        'Respond to @mentions, keywords, and AI relevance (default).',
+    },
+    {
+      name: 'keyword',
+      description: 'Respond to @mentions and keyword matches only.',
+    },
+    {
+      name: 'none',
+      description: 'Never respond in this channel. (channel only)',
+    },
+  ],
+};
+
 import { handleClear } from './clear';
 import { handleSet } from './set';
 import { showMode } from './show';
